+++ src/main/java/chess/domain/board/ChessBoardFactory.java
@@ -5,14 +5,14 @@ import chess.domain.piece.Piece;
 import chess.domain.piece.position.File;
 import chess.domain.piece.position.PiecePosition;
 import chess.domain.piece.position.Rank;
-import chess.domain.piece.strategy.BishopMovementStrategy;
-import chess.domain.piece.strategy.KingMovementStrategy;
-import chess.domain.piece.strategy.KnightMovementStrategy;
-import chess.domain.piece.strategy.QueenMovementStrategy;
-import chess.domain.piece.strategy.RookMovementStrategy;
-import chess.domain.piece.strategy.pawn.BlackPawnMovementStrategy;
-import chess.domain.piece.strategy.pawn.PawnMovementStrategy;
-import chess.domain.piece.strategy.pawn.WhitePawnMovementStrategy;
+import chess.domain.piece.movestrategy.BishopMovementStrategy;
+import chess.domain.piece.movestrategy.KingMovementStrategy;
+import chess.domain.piece.movestrategy.KnightMovementStrategy;
+import chess.domain.piece.movestrategy.QueenMovementStrategy;
+import chess.domain.piece.movestrategy.RookMovementStrategy;
+import chess.domain.piece.movestrategy.pawn.BlackPawnMovementStrategy;
+import chess.domain.piece.movestrategy.pawn.PawnMovementStrategy;
+import chess.domain.piece.movestrategy.pawn.WhitePawnMovementStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
