+++ src/main/java/chess/domain/piece/movestrategy/KingMovementStrategy.java
@@ -1,14 +1,14 @@
-package chess.domain.piece.strategy;
+package chess.domain.piece.movestrategy;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.position.PiecePosition;
 
-public class QueenMovementStrategy extends AbstractPieceMovementStrategy {
+public class KingMovementStrategy extends AbstractPieceMovementStrategy {
 
-    private static final double QUEEN_VALUE = 9;
+    private static final double KING_VALUE = 0;
 
-    public QueenMovementStrategy(final Color color) {
+    public KingMovementStrategy(final Color color) {
         super(color);
     }
 
@@ -16,13 +16,18 @@ public class QueenMovementStrategy extends AbstractPieceMovementStrategy {
     protected void validateMoveWithNoAlly(final PiecePosition source,
                                           final PiecePosition destination,
                                           final Piece nullableEnemy) throws IllegalArgumentException {
-        if (!isStraight(source, destination) && !isDiagonal(source, destination)) {
-            throw new IllegalArgumentException("퀸은 대각선 혹은 직선으로만 이동가능합니다.");
+        if (!isUnitDistance(source, destination)) {
+            throw new IllegalArgumentException("왕은 한칸만 이동할 수 있습니다.");
         }
     }
 
     @Override
     public double judgeValue() {
-        return QUEEN_VALUE;
+        return KING_VALUE;
+    }
+
+    @Override
+    public boolean isKing() {
+        return true;
     }
 }
