+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -4,14 +4,14 @@ import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.position.PiecePosition;
 import chess.domain.piece.position.Rank;
-import chess.domain.piece.strategy.BishopMovementStrategy;
-import chess.domain.piece.strategy.KingMovementStrategy;
-import chess.domain.piece.strategy.KnightMovementStrategy;
-import chess.domain.piece.strategy.QueenMovementStrategy;
-import chess.domain.piece.strategy.RookMovementStrategy;
-import chess.domain.piece.strategy.pawn.BlackPawnMovementStrategy;
-import chess.domain.piece.strategy.pawn.PawnMovementStrategy;
-import chess.domain.piece.strategy.pawn.WhitePawnMovementStrategy;
+import chess.domain.piece.movestrategy.BishopMovementStrategy;
+import chess.domain.piece.movestrategy.KingMovementStrategy;
+import chess.domain.piece.movestrategy.KnightMovementStrategy;
+import chess.domain.piece.movestrategy.QueenMovementStrategy;
+import chess.domain.piece.movestrategy.RookMovementStrategy;
+import chess.domain.piece.movestrategy.pawn.BlackPawnMovementStrategy;
+import chess.domain.piece.movestrategy.pawn.PawnMovementStrategy;
+import chess.domain.piece.movestrategy.pawn.WhitePawnMovementStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
