+++ src/test/java/chess/domain/game/MovePieceTest.java
@@ -1,6 +1,8 @@
 package chess.domain.game;
 
 import chess.domain.board.ChessBoardFactory;
+import chess.domain.game.state.ChessGameState;
+import chess.domain.game.state.InitializeGame;
 import chess.domain.piece.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -21,11 +23,11 @@ class MovePieceTest {
     @Test
     void 폰을_움직인_후_차례를_바꾼다() {
         // given
-        ChessGameStep step = new InitializeGame(new ChessBoardFactory().create());
+        ChessGameState step = new InitializeGame(new ChessBoardFactory().create());
         step = step.initialize();
 
         // when
-        final ChessGameStep next = step.movePiece(of("e2"), of("e4"));
+        final ChessGameState next = step.movePiece(of("e2"), of("e4"));
 
         // then
         assertAll(
@@ -41,17 +43,17 @@ class MovePieceTest {
     @Test
     void 왕을_죽인_경우_승자를_EndGame_에_전달한다() {
         // given
-        ChessGameStep step = new InitializeGame(new ChessBoardFactory().create());
+        ChessGameState step = new InitializeGame(new ChessBoardFactory().create());
         step = step.initialize();
 
         // when
-        final ChessGameStep next = 흰색_왕을_죽인다(step);
+        final ChessGameState next = 흰색_왕을_죽인다(step);
 
         // then
         assertThat(next.winColor()).isEqualTo(Color.BLACK);
     }
 
-    private ChessGameStep 흰색_왕을_죽인다(ChessGameStep step) {
+    private ChessGameState 흰색_왕을_죽인다(ChessGameState step) {
         step = step.movePiece(of("e2"), of("e4"));
         step = step.movePiece(of("d7"), of("d5"));
         step = step.movePiece(of("e1"), of("e2"));
@@ -64,11 +66,11 @@ class MovePieceTest {
     @Test
     void 진행중인_상태에서_end_실행_시_무승부_상태로_EndGame_으로_넘어간다() {
         // given
-        ChessGameStep step = new InitializeGame(new ChessBoardFactory().create());
+        ChessGameState step = new InitializeGame(new ChessBoardFactory().create());
         step = step.initialize();
 
         // when
-        final ChessGameStep next = step.end();
+        final ChessGameState next = step.end();
 
         // then
         assertThat(next.winColor()).isEqualTo(Color.NONE);
