+++ src/test/java/chess/domain/piece/movestrategy/QueenMovementTest.java
@@ -1,4 +1,4 @@
-package chess.domain.piece.strategy;
+package chess.domain.piece.movestrategy;
 
 import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
@@ -15,39 +15,51 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
 
 import static chess.domain.piece.position.PiecePositionFixture.piecePositions;
-import static java.util.Collections.emptyList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@DisplayName("BishopMovementStrategy 은")
-class BishopMovementStrategyTest {
+@DisplayName("QueenMovementStrategy 은")
+class QueenMovementTest {
 
     private final Color myColor = Color.WHITE;
     private final Color enemyColor = Color.BLACK;
-    private final PieceMovementStrategy movement = new BishopMovementStrategy(myColor);
+    private final PieceMovementStrategy movement = new QueenMovementStrategy(myColor);
     private final PiecePosition source = PiecePosition.of("e4");
 
     @Nested
     @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class 대각선_경로로_움직이는_경우 {
+    class 직선_혹은_대각선_경로로_움직이는_경우 {
 
         @ParameterizedTest(name = "움직일 수 있다. [e4] -> [{1}]")
         @CsvSource({
-                "d3",
-                "d5",
-                "f3",
                 "f5",
-                "c2",
-                "c6",
-                "g2",
+                "f3",
+                "d5",
+                "d3",
                 "g6",
+                "g2",
+                "c6",
+                "c2",
+                "f4",
+                "d4",
+                "e5",
+                "e3",
+                "g4",
+                "c4",
+                "e6",
+                "e2",
+                "h4",
+                "a4",
+                "e8",
+                "e1",
         })
         void 움직일_수_있다(final PiecePosition destination) {
             // when & then
@@ -55,35 +67,47 @@ class BishopMovementStrategyTest {
         }
 
         @ParameterizedTest(name = "경유지를 반환한다. 출발: [e4] -> 경유지: [{1}] -> 도착: [{0}]")
-        @MethodSource("bishopDestinations")
+        @MethodSource("queenDestinations")
         void 경유지를_반환한다(final PiecePosition destination, final List<PiecePosition> waypoints) {
             // when & then
             assertThat(movement.waypoints(source, destination, null)).containsExactlyInAnyOrderElementsOf(waypoints);
         }
 
-        Stream<Arguments> bishopDestinations() {
+        Stream<Arguments> queenDestinations() {
             return Stream.of(
-                    Arguments.of("f5", Named.of("없다", emptyList())),
-                    Arguments.of("f3", Named.of("없다", emptyList())),
-                    Arguments.of("d5", Named.of("없다", emptyList())),
-                    Arguments.of("d3", Named.of("없다", emptyList())),
+                    Arguments.of("f5", Named.of("없다", Collections.emptyList())),
+                    Arguments.of("f3", Named.of("없다", Collections.emptyList())),
+                    Arguments.of("d5", Named.of("없다", Collections.emptyList())),
+                    Arguments.of("d3", Named.of("없다", Collections.emptyList())),
                     Arguments.of("g6", Named.of("f5", piecePositions("f5"))),
                     Arguments.of("g2", Named.of("f3", piecePositions("f3"))),
                     Arguments.of("c6", Named.of("d5", piecePositions("d5"))),
-                    Arguments.of("c2", Named.of("d3", piecePositions("d3")))
+                    Arguments.of("c2", Named.of("d3", piecePositions("d3"))),
+                    Arguments.of("f4", Named.of("없다", Collections.emptyList())),
+                    Arguments.of("d4", Named.of("없다", Collections.emptyList())),
+                    Arguments.of("e5", Named.of("없다", Collections.emptyList())),
+                    Arguments.of("e3", Named.of("없다", Collections.emptyList())),
+                    Arguments.of("g4", Named.of("f4", piecePositions("f4"))),
+                    Arguments.of("c4", Named.of("d4", piecePositions("d4"))),
+                    Arguments.of("e6", Named.of("e5", piecePositions("e5"))),
+                    Arguments.of("e2", Named.of("e3", piecePositions("e3"))),
+                    Arguments.of("h4", Named.of("f4, g4", piecePositions("f4", "g4"))),
+                    Arguments.of("a4", Named.of("d4, c4, b4", piecePositions("d4", "c4", "b4"))),
+                    Arguments.of("e8", Named.of("e5, e6, e7", piecePositions("e5", "e6", "e7"))),
+                    Arguments.of("e1", Named.of("e2, e3", piecePositions("e2", "e3")))
             );
         }
     }
 
     @Nested
-    class 대각선이_아닌_경로로_움직이는_경우 {
+    class 직선_혹은_대각선이_아닌_경로로_움직이는_경우 {
 
         @ParameterizedTest(name = "움직일 수 없다. [e4] -> [{1}]")
         @CsvSource({
-                "e5",
-                "e3",
-                "d4",
-                "f4",
+                "g5",
+                "c5",
+                "h6",
+                "a2",
         })
         void 움직일_수_없다(final PiecePosition destination) {
             // when & then
@@ -93,10 +117,10 @@ class BishopMovementStrategyTest {
 
         @ParameterizedTest(name = "경유지를 조회하면 예외. [e4] -> [{1}]")
         @CsvSource({
-                "e5",
-                "e3",
-                "d4",
-                "f4",
+                "g5",
+                "c5",
+                "h6",
+                "a2",
         })
         void 경유지를_조회하면_예외(final PiecePosition destination) {
             // when & then
@@ -108,7 +132,7 @@ class BishopMovementStrategyTest {
     @Test
     void 아군을_죽일_수_없다() {
         // given
-        final PiecePosition dest = PiecePosition.of("d3");
+        final PiecePosition dest = PiecePosition.of("e6");
         final Piece ally = new Piece(dest, new RookMovementStrategy(myColor));
 
         // when & then
@@ -119,7 +143,7 @@ class BishopMovementStrategyTest {
     @Test
     void 적군을_죽일_수_있다() {
         // given
-        final PiecePosition dest = PiecePosition.of("d3");
+        final PiecePosition dest = PiecePosition.of("e6");
         final Piece enemy = new Piece(dest, new RookMovementStrategy(enemyColor));
 
         // when & then
